Created Food List View

+++ app/src/main/java/com/mobdeve/s17/catchow/models/Rating.java
@@ -2,16 +2,16 @@ package com.mobdeve.s17.catchow.models;
 
 public class Rating {
     private String name;
-    private Integer taste;
-    private Integer packaging;
-    private Integer price;
+    private long taste;
+    private long packaging;
+    private long price;
     private String review;
 
     public Rating() {
         // Default constructor
     }
 
-    public Rating(String name, Integer taste, Integer packaging, Integer price, String review) {
+    public Rating(String name, long taste, long packaging, long price, String review) {
         this.name = name;
         this.taste = taste;
         this.packaging = packaging;
@@ -27,27 +27,27 @@ public class Rating {
         this.name = name;
     }
 
-    public Integer getTaste() {
+    public long getTaste() {
         return taste;
     }
 
-    public void setTaste(Integer taste) {
+    public void setTaste(long taste) {
         this.taste = taste;
     }
 
-    public Integer getPackaging() {
+    public long getPackaging() {
         return packaging;
     }
 
-    public void setPackaging(Integer packaging) {
+    public void setPackaging(long packaging) {
         this.packaging = packaging;
     }
 
-    public Integer getPrice() {
+    public long getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(long price) {
         this.price = price;
     }
